feat: endpoint users

+++ src/main/java/gia/utp/primerproyecto/primerproyecto/model/repository/HotelRepository.java
@@ -3,8 +3,7 @@ package gia.utp.primerproyecto.primerproyecto.model.repository;
 import gia.utp.primerproyecto.primerproyecto.model.entities.HotelEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 
-public interface HotelRepository extends JpaRepository<HotelEntity, Integer>{
+public interface HotelRepository extends JpaRepository<HotelEntity, Long>{
 
 }
